Move add_created_profile_to_configuration test
