+++ crates/cast/src/starknet_commands/account/mod.rs
@@ -156,3 +156,44 @@ pub fn add_created_profile_to_configuration(
 
     Ok(())
 }
+
+#[cfg(test)]
+mod tests {
+    use cast::helpers::constants::DEFAULT_ACCOUNTS_FILE;
+    use cast::helpers::scarb_utils::CastConfig;
+    use sealed_test::prelude::rusty_fork_test;
+    use sealed_test::prelude::sealed_test;
+    use std::fs;
+
+    use crate::starknet_commands::account::add_created_profile_to_configuration;
+
+    #[sealed_test(files = ["tests/data/contracts/v1/balance/Scarb.toml"])]
+    fn test_add_created_profile_to_configuration_happy_case() {
+        let config = CastConfig {
+            rpc_url: String::from("http://some-url"),
+            account: String::from("some-name"),
+            accounts_file: "accounts".into(),
+        };
+        let res = add_created_profile_to_configuration(&None, &config);
+
+        assert!(res.is_ok());
+
+        let contents = fs::read_to_string("Scarb.toml").expect("Unable to read Scarb.toml");
+        assert!(contents.contains("[tool.sncast.some-name]"));
+        assert!(contents.contains("account = \"some-name\""));
+        assert!(contents.contains("url = \"http://some-url\""));
+        assert!(contents.contains("accounts-file = \"accounts\""));
+    }
+
+    #[sealed_test(files = ["tests/data/contracts/v1/balance/Scarb.toml"])]
+    fn test_add_created_profile_to_configuration_profile_already_exists() {
+        let config = CastConfig {
+            rpc_url: String::from("http://some-url"),
+            account: String::from("myprofile"),
+            accounts_file: DEFAULT_ACCOUNTS_FILE.into(),
+        };
+        let res = add_created_profile_to_configuration(&None, &config);
+
+        assert!(res.is_err());
+    }
+}
